+++ pyincore_viz/geoutil.py
@@ -1473,3 +1473,72 @@ class GeoUtil:
         # # fill empty value as blank
 
         return choro_data
+
+    @staticmethod
+    def plot_hazard(dataset_id, client):
+        """Plot hazard dataset on the map
+
+        args:
+            hazard_dataset (obj): pyincore HazardDataset object
+            dataset_id (str): ID of tornado hazard.
+            client (obj): pyincore service Client Object.
+
+        returns:
+            none
+        """
+        metadata = DataService(client).get_dataset_metadata(dataset_id)
+        # metadata = DataService(client)
+        title = metadata['title']
+
+        dataset = Dataset.from_data_service(dataset_id, DataService(client))
+        raster_file_path = Path(dataset.local_file_path). \
+            joinpath(dataset.metadata['fileDescriptors'][0]['filename'])
+
+        GeoUtil.plot_raster_file_with_legend(raster_file_path, title)
+        hazard_type = hazard_dataset.hazard_type
+        title = "Hazard Type: " + hazard_type
+        raster_file_path = hazard_dataset.dataset.local_file_path
+
+        GeoUtil.plot_raster_file_with_legend(raster_file_path, title)
+
+    @staticmethod
+    def plot_local_hazard(dataset):
+        """Plot hazard dataset on the map
+
+        args:
+            dataset (obj): pyincore HazardDataset object
+
+        returns:
+            none
+        """
+        hazard_type = dataset.hazard_type
+
+        if hazard_type.lower() == "earthquake":
+            if len(dataset.hazardDatasets) > 1:
+                for earthquake in dataset.hazardDatasets:
+                    GeoUtil.plot_local_earthquake(earthquake)
+            else:
+                GeoUtil.plot_local_earthquake(dataset.hazardDatasets[0])
+        elif hazard_type.lower() == "tsunami":
+            if len(dataset.hazardDatasets) > 1:
+                for tsunami in dataset.hazardDatasets:
+                    GeoUtil.plot_local_tsunami(tsunami)
+            else:
+                GeoUtil.plot_local_tsunami(dataset.hazardDataset[0])
+        elif hazard_type.lower() == "flood":
+            if len(dataset.hazardDatasets) > 1:
+                for flood in dataset.hazardDatasets:
+                    GeoUtil.plot_local_flood(flood)
+            else:
+                GeoUtil.plot_local_flood(dataset.hazardDatasets[0])
+        elif hazard_type.lower() == "hurricane":
+            if len(dataset.hazardDatasets) > 1:
+                for hurricane in dataset.hazardDatasets:
+                    GeoUtil.plot_local_hurricane(hurricane)
+            else:
+               GeoUtil.plot_local_hurricane(dataset.hazardDatasets[0])
+        elif hazard_type.lower() == "tornado":
+            print("The given hazard type is not supported because tornado is a shapefile.")
+            print("Please use plot_local_tornado method instead.")
+
+        title = "Hazard Type: " + hazard_type